Store times as epoch seconds rather than date strings

+++ src/leaderboard/chrono/real_time_provider.py
@@ -8,6 +8,6 @@ from src.leaderboard.chrono.time_provider import TimeProvider
 class RealTimeProvider(TimeProvider):
   """Provides the current time."""
 
-  def get_current_time(self) -> float:
+  def get_current_time(self) -> int:
     """Return the current time."""
-    return time.time()
+    return int(time.time())
